Hecho: 1. Creado el sistema de transiciones
Pendiente:

1. Crear los ríos y el agua, en general.
2. Introducción de nuevos eventos.
3. Hacer que los eventos caduquen pasado el tiempo y que haya
consecuencias.
4. Hacer que los eventos, en lugar de aparecer repentinamente, se añadan
a una cola desplegable.
5. Revisión de criterios de cumplimiento antes y después de que un
evento haya sido añadido. Actualización de opciones según estado de las
variables.
6. En "Opción" 64:9 buscar una solución más elegante (Depende del
punto anterior).
7. Producción:
Deben crearse los "edificios productivos", las granjas, los registros de
granjas y edificios productivos, fijar las producciones máximas, las
localizaciones,los individuos que trabajan en ellas y cansarlos conforme
se vayan "usando". También hay que hacer que el individuo pueda subir de
nivel. La familia necesita un inventario, también la granja en la que
estén trabajando.
8. Hacer las guías.
9. Seguir con el registro de funciones.
10. En Registrodefunciones\instalacióndenuevosciudadanos hacer que la
edad y quizá otras variables sean aleatorias.
11. Cambiar la perspectiva del juego.
12. Mirar la clase EventosyMenú y plantearse su utilidad.
13. Llenar y reordenar el resto de menús.
14. Creación de familias y especialización de las familias. Añadir
niveles a los edificios.
15. Eliminar las referencias a las texturas que no se usen.

+++ Ciudadanos/core/src/com/virtualgames/ciudadanos/Objetodejuego.java
@@ -5,8 +5,8 @@ import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 public abstract class Objetodejuego {
 	public float posicionX;
 	public float posicionY;
-	public int dimensionX;
-	public int dimensionY;
+	public float dimensionX;
+	public float dimensionY;
 	public int origenX;
 	public int origenY;
 	public int escalaX;
